fix(CAPTCHA): centralize JPA id-based equals and hashCode in BaseEntity
